Subscriber position persistence
git-tfs-id: [https://chrispday.visualstudio.com:443/DefaultCollection]$/EventStore;C34

+++ Yeast/Yeast.EventStore/Provider/SQL/SqlServerEventStoreProviderPosition.cs
@@ -7,7 +7,7 @@ namespace Yeast.EventStore.Provider
 {
 	public class SqlServerEventStoreProviderPosition : IEventStoreProviderPosition
 	{
-		public int Position;
+		public long Position;
 
 		public override string ToString()
 		{
